+++ src/services/dayjs-service.ts
@@ -1,0 +1,12 @@
+// This is the one place we import & configure dayjs
+// eslint-disable-next-line unicorn/import-style
+import dayjsBase from 'dayjs'
+import pluralGetSet from 'dayjs/plugin/pluralGetSet'
+import quarterOfYear from 'dayjs/plugin/quarterOfYear'
+import utc from 'dayjs/plugin/utc'
+
+dayjsBase.extend(utc)
+dayjsBase.extend(quarterOfYear)
+dayjsBase.extend(pluralGetSet)
+
+export const dayjs = dayjsBase