+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * **features:** Table - list species name in the table (#82)
 * **features:** Export species richness raw data as csv (#36)
 * **features:** Activity patterns species selector (#89)
+* **features:** Detected species by time (#20)
 * **features:** Each species row in table should be clickable and link to ap page (#110)
 
-* **refactor:** Refactor species richness tabledisplay (#112)
+* **refactor:** Refactor species richness table display (#112)